+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/junit/ExampleDomainConverters.java
@@ -1,14 +1,14 @@
 package io.github.nchaugen.tabletest.junit;
 
-import io.github.nchaugen.tabletest.junit.exampledomain.FactoryMethodDate;
+import io.github.nchaugen.tabletest.junit.exampledomain.ExternalFactoryDate;
 
 import java.time.LocalDate;
 
 @SuppressWarnings("unused")
 public class ExampleDomainConverters {
 
-    public static FactoryMethodDate parseDate(String date) {
-        return new FactoryMethodDate(LocalDate.parse(date));
+    public static ExternalFactoryDate parseDate(String date) {
+        return new ExternalFactoryDate(LocalDate.parse(date));
     }
 
 }