+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/junit/exampledomain/ExternalFactoryDate.java
@@ -2,4 +2,4 @@ package io.github.nchaugen.tabletest.junit.exampledomain;
 
 import java.time.LocalDate;
 
-public record FactoryMethodDate(LocalDate date) {}
+public record ExternalFactoryDate(LocalDate date) {}